==> notebooks/02. Building Recommender of Movie Lens dataset.ipynb <==
# Analysis of MovieLens dataset (Beginner'sAnalysis)
https://www.kaggle.com/jneupane12/analysis-of-movielens-dataset-beginner-sanalysis

import sys
print(sys.executable)

## 1. First we import necessary Libaries

import pandas as pd # pandas is a data manipulation library
import numpy as np #provides numerical arrays and functions to manipulate the arrays efficiently
import random
import matplotlib.pyplot as plt # data visualization library
import wordcloud #used to generate world cloud
import time
import datetime
import re

## 2.Reading and Exploring the Data

### 2.1 Load Movies Data

movies = pd.read_csv("datasets/movies.csv")

movies.sample(5)

movies.shape, movies.info(), len(movies['movieId'].unique().tolist())

### 2.2 Load Ratings Data

ratings = pd.read_csv("datasets/ratings.csv")
ratings.info()

ratings.sample(5)

ratings.describe()

ratings.shape, len(ratings['movieId'].unique().tolist())

### 2.3 Load Tags Data

tags = pd.read_csv("datasets/tags.csv")

ratings.shape, len(ratings['movieId'].unique().tolist())

tags.sample(5)

## 3.Cleaning of data

movies.isnull().any()

ratings.isnull().any()

tags.isnull().any()

# # lets drop null rows
# tags=tags.dropna()

## 4.Data Analysis

### 4.1 Make dates more readable

# https://stackoverflow.com/a/62968313/2049763
def timestamp_to_date_converter(t):
    # https://realpython.com/python-time-module/
    return datetime.datetime.fromtimestamp(t).strftime("%A, %B %d, %Y %I:%M:%S")

ratings['date'] = ratings['timestamp'].apply(timestamp_to_date_converter)
tags['date']    = tags['timestamp'].apply(timestamp_to_date_converter)

ratings.sort_values(by=['timestamp'], ascending=True).tail(5)

s = "01/01/2018"
t = time.mktime(datetime.datetime.strptime(s, "%m/%d/%Y").timetuple())

ratings.loc[ratings.timestamp>t].sort_values(by=['timestamp'], ascending=True).head(5)

ratings.loc[ratings.timestamp>t].sort_values(by=['timestamp'], ascending=True).shape

### 4.2 Reading Movie Release Year

# https://stackoverflow.com/a/8569258
def title_to_release_year(s):
    m = re.findall(r"\(([0-9]+)\)", s)
    
    # https://www.guru99.com/python-regular-expressions-complete-tutorial.html
    if m is None or len(m) <= 0:
        return None        
    return m[-1]

movies['year']  = movies['title'].apply(title_to_release_year)

movies.sample(5)

movies.isnull().any()

# https://datatofish.com/rows-with-nan-pandas-dataframe/
movies[movies.isnull().any(axis=1)]

movies = movies.dropna()

## 5. Basic Recomendation 

https://www.kaggle.com/subhamoybhaduri/diff-approaches-of-building-recommender-system

ratings.shape

# https://stackoverflow.com/a/39881230
# ratings =  ratings.loc[ratings.movieId.isin(movies.movieId)] 

movies_ratings = ratings.merge(movies, on = 'movieId', how = 'inner')
movies_ratings.shape

movies_ratings = movies_ratings.sort_values(['year', 'timestamp'], ascending=[True, True])
movies_ratings.head(5)

nb_users  = movies_ratings['userId'].nunique()
nb_movies = movies_ratings['movieId'].nunique()

nb_users, nb_movies 

ratings_matrix = movies_ratings.pivot_table(index=['userId'],columns=['movieId'],values='rating').reset_index(drop=True)
ratings_matrix.fillna(0, inplace = True)

ratings_matrix.sample(5)

data_matrix = np.array(ratings_matrix)
print(data_matrix.shape)

### 5.1 Gaussian Mixture Model and Expectation-Maximization Algorithm

from sklearn.model_selection import train_test_split

from sklearn.mixture import GaussianMixture
from scipy.special import logsumexp
import itertools

# split the data into train and test set
train, test = train_test_split(data_matrix, test_size=0.2, random_state=42, shuffle=True)

test.shape, test

# https://jakevdp.github.io/PythonDataScienceHandbook/05.12-gaussian-mixtures.html#How-many-components?
gmm_model = GaussianMixture(n_components=2, covariance_type='full', 
                            tol=0.001, reg_covar=1e-06, max_iter=100, 
                            n_init=1, init_params='kmeans', weights_init=None, 
                            means_init=None, precisions_init=None, random_state=42, 
                            warm_start=False, verbose=0, verbose_interval=10)
gmm_model.fit(train)

print(gmm_model.means_.shape)
print(gmm_model.covariances_.shape)
print(gmm_model.weights_.shape)

gmm_model.predict(test)

# Fill Missing Values i.e Recommend
inver0, inver1 = gmm_model.covariances_[0], gmm_model.covariances_[1] 
inver0, inver1 = np.linalg.inv(inver0), np.linalg.inv(inver1)


deter0, deter1 = gmm_model.covariances_[0], gmm_model.covariances_[1] 
deter0, deter1 = np.linalg.det(deter0), np.linalg.det(deter1)

n, d = train.shape
K = gmm_model.means_.shape[0]
print(n, d, K)

mean = gmm_model.means_
variance = gmm_model.covariances_
weight = np.log(gmm_model.weights_)

calc = np.zeros((n, K))
ind = np.zeros((n, d))

soft = calc
X_pred = ind

add = np.zeros((n,))
dim = np.zeros((n,))
    
ind = np.where(train != 0, 1, 0)            
dim = np.sum(ind, axis=1)

for i in range(n):
    for j in range(K):
        res = train[i] - mean[j]
        res = np.multiply(res, ind[i])
        
        # Multivariate Gaussian
        if j == 0:
            A = (res.T @ inver0) @ res
            C = (dim[i]/2)*np.log(2*np.pi) + np.log(deter0 + 1e-16)/2
        else:
            A = (res.T @ inver1) @ res
            C = (dim[i]/2)*np.log(2*np.pi) + np.log(deter1 + 1e-16)/2
        
        B = 2
        calc[i, j] = weight[j] + (-A/B) - C

add = logsumexp(calc, axis = 1)

#Since the entire computation is done in log-domain to avoid Numerical instability
#we need to bring it back in its original domain
soft = np.exp(np.subtract(np.transpose(calc), add))

lg = np.sum(add)
    
X_calc = np.transpose(soft) @ gmm_model.means_

#We will use predicted value if the entry is 0 in original rating matrix
data_matrix_pred_GMM = np.where(data_matrix == 0, X_calc, data_matrix)

for i in range(data_matrix_pred_GMM.shape[0]):
    for j in range(data_matrix_pred_GMM.shape[1]):
        data_matrix_pred_GMM[i, j] = round(data_matrix_pred_GMM[i, j])

#For measuring the performance we have to use the predicted matrix
for i in range(X_calc.shape[0]):
    for j in range(X_calc.shape[1]):
        X_pred[i, j] = round(X_calc[i, j])

ind_matrix = np.zeros((nb_users, nb_movies))
ind_matrix = np.where(data_matrix != 0, 1, 0)

x = np.multiply(X_pred, ind_matrix)

RMSE_GMM = np.sqrt(np.mean((x - data_matrix)**2))
print("RMSE of GMM Model is %f." %RMSE_GMM)

### 5.2 Altenating Least Squares using Non-Negative Matrix Factorization
