==> notebooks/04. Latent Matrix Factorization Recommender.ipynb <==
# Analysis of MovieLens dataset (Beginner'sAnalysis)

Data Source: https://www.kaggle.com/jneupane12/analysis-of-movielens-dataset-beginner-sanalysis
<br>Algorithm: https://towardsdatascience.com/introduction-to-latent-matrix-factorization-recommender-systems-8dfc63b94875
<br>Notebook: https://www.kaggle.com/robottums/probabalistic-matrix-factorization-with-suprise

import sys
print(sys.executable)

import os
os.environ['PATH']

## 1. First we import necessary Libaries

import numpy as np #provides numerical arrays and functions to manipulate the arrays efficiently
import pandas as pd # pandas is a data manipulation library
import random
import matplotlib.pyplot as plt # data visualization library
import time
import datetime
import re

## 2.Reading and Exploring the Data

movies = pd.read_csv("datasets/movies.csv")
ratings = pd.read_csv("datasets/ratings.csv")
tags = pd.read_csv("datasets/tags.csv")

### 2.1 Load Movies Data

movies.sample(5)

movies.shape, movies.info(), len(movies['movieId'].unique().tolist())

### 2.2 Load Ratings Data

ratings.drop_duplicates(inplace=True)
print('we have',ratings.shape[0], 'ratings')
print('the number of unique users we have is:', len(ratings.userId.unique()))
print('the number of unique movies we have is:', len(ratings.movieId.unique()))
print("The median user rated %d books."%ratings.userId.value_counts().median())
print('The max rating is: %d'%ratings.rating.max(),"the min rating is: %d"%ratings.rating.min())

ratings.sample(5)

ratings.shape, "\n", ratings.info(), "\n", len(ratings['movieId'].unique().tolist()), "\n", len(ratings['userId'].unique().tolist())
# ratings.describe()

### 2.3 Load Tags Data

tags.sample(5)

tags.shape, "\n", tags.info(), len(tags['tag'].unique().tolist())

## 3.Cleaning of data

movies.isnull().any(), " ", ratings.isnull().any(), " ", tags.isnull().any()

# # lets drop null rows
# tags=tags.dropna()

## 4.Data Analysis

### 4.1 Make dates more readable

# https://stackoverflow.com/a/62968313/2049763
def timestamp_to_date_converter(t):
    # https://realpython.com/python-time-module/
    return datetime.datetime.fromtimestamp(t).strftime("%A, %B %d, %Y %I:%M:%S")

ratings['date'] = ratings['timestamp'].apply(timestamp_to_date_converter)
tags['date']    = tags['timestamp'].apply(timestamp_to_date_converter)

ratings.sort_values(by=['timestamp'], ascending=True).tail(5)

s = "01/01/2018"
t = time.mktime(datetime.datetime.strptime(s, "%m/%d/%Y").timetuple())

ratings.loc[ratings.timestamp>t].sort_values(by=['timestamp'], ascending=True).head(5)

ratings.loc[ratings.timestamp>t].sort_values(by=['timestamp'], ascending=True).shape

### 4.2 Reading Movie Release Year

# https://stackoverflow.com/a/8569258
def title_to_release_year(s):
    m = re.findall(r"\(([0-9]+)\)", s)
    
    # https://www.guru99.com/python-regular-expressions-complete-tutorial.html
    if m is None or len(m) <= 0:
        return None        
    return m[-1]

movies['year']  = movies['title'].apply(title_to_release_year)

movies.sample(5)

movies.isnull().any()

# https://datatofish.com/rows-with-nan-pandas-dataframe/
movies[movies.isnull().any(axis=1)]

movies = movies.dropna()

### 4.3 Preparing ratings data

ratings.userId = ratings.userId.astype('category').cat.codes.values
ratings.movieId = ratings.movieId.astype('category').cat.codes.values

ratings.info(), ratings['userId'].value_counts(ascending=True)

## 5. Basic Recomendation 

https://www.kaggle.com/rajmehra03/cf-based-recsys-by-low-rank-matrix-factorization

# https://stackoverflow.com/a/39881230
# ratings =  ratings.loc[ratings.movieId.isin(movies.movieId)] 

movies_ratings = ratings.merge(movies, on = 'movieId', how = 'inner')
movies_ratings.shape, '# ============== #', ratings.isnull().any()

movies_ratings = movies_ratings.sort_values(['year', 'timestamp'], ascending=[True, True])
movies_ratings.head(5)

### 5.1 Data loading using Surprise

import surprise

#swapping columns
raw=movies_ratings[['userId','movieId','rating']] 
# when importing from a DF, you only need to specify the scale of the ratings.
reader = surprise.Reader(rating_scale=(1,5)) 
#into surprise:
data_movies_ratings = surprise.Dataset.load_from_df(raw,reader)

### 5.2 Model

class ProbabilisticMatrixFactorization(surprise.AlgoBase):
# Randomly initializes two Matrices, Stochastic Gradient Descent to be able to optimize the best factorization for ratings.
    def __init__(self,learning_rate,num_epochs,num_factors):
       # super(surprise.AlgoBase)
        self.alpha = learning_rate #learning rate for Stochastic Gradient Descent
        self.num_epochs = num_epochs
        self.num_factors = num_factors
    
    def fit(self,train):
        #randomly initialize user/item factors from a Gaussian
        P = np.random.normal(0,.1,(train.n_users,self.num_factors))
        Q = np.random.normal(0,.1,(train.n_items,self.num_factors))
        #print('fit')

        for epoch in range(self.num_epochs):
            for u,i,r_ui in train.all_ratings():
                residual = r_ui - np.dot(P[u],Q[i])
                temp = P[u,:] # we want to update them at the same time, so we make a temporary variable. 
                P[u,:] +=  self.alpha * residual * Q[i]
                Q[i,:] +=  self.alpha * residual * temp 

                
        self.P = P
        self.Q = Q

        self.trainset = train
    
    
    def estimate(self,u,i):
        # returns estimated rating for user u and item i. 
        # Prerequisite: Algorithm must be fit to training set.
        # check to see if u and i are in the train set:
        # print('gahh')

        if self.trainset.knows_user(u) and self.trainset.knows_item(i):
            #print(u,i, '\n','yep:', self.P[u],self.Q[i])
            #return scalar product of P[u] and Q[i]
            nanCheck = np.dot(self.P[u],self.Q[i])
            
            if np.isnan(nanCheck):
                print('Got NaN error, returning the general average')
                return self.trainset.global_mean
            else:
                return np.dot(self.P[u,:],self.Q[i,:])
        else: # if its not known we'll return the general average. 
            print('its not known, returning the general average')
            return self.trainset.global_mean

gs = surprise.model_selection.GridSearchCV(ProbabilisticMatrixFactorization, 
                                           param_grid={'learning_rate':[0.005,0.01],
                                                       'num_epochs':[5,10],
                                                       'num_factors':[10,20]},
                                           measures=['rmse', 'mae'], cv=2)

gs.fit(data_movies_ratings)

print('rsme:', gs.best_score['rmse'],'mae:',gs.best_score['mae'])

best_params = gs.best_params['rmse']
print('best_params \n{\n  rsme: ', gs.best_params['rmse'], '\n  mae:  ', gs.best_params['mae'], " \n}")

### 5.2.1 Evaluating the Model Performance

bestVersion = ProbabilisticMatrixFactorization(learning_rate=best_params['learning_rate'],
                                               num_epochs=best_params['num_epochs'],
                                               num_factors=best_params['num_factors'])

#we can use k-fold cross validation to evaluate the best model. 
kSplit = surprise.model_selection.KFold(n_splits=10,shuffle=True)
for train,test in kSplit.split(data_movies_ratings):
    bestVersion.fit(train)
    prediction = bestVersion.test(test)
    surprise.accuracy.rmse(prediction,verbose=True)

### 5.2.2 Evaluating the Model Performance

Alg1 = ProbabilisticMatrixFactorization(learning_rate=best_params['learning_rate'],
                                        num_epochs=best_params['num_epochs'],
                                        num_factors=best_params['num_factors'])

data1 = data_movies_ratings.build_full_trainset()
Alg1.fit(data1)

for i in range(10):
    index = random.randint(1, movies_ratings.shape[0] - 1)
    print(index, ": ", movies_ratings.userId.iloc[index], movies_ratings.movieId.iloc[index], 
          movies_ratings.title.iloc[index], ":", 
          round(Alg1.estimate(movies_ratings.userId.iloc[index],movies_ratings.movieId.iloc[index])), 
          " <---> ", round(movies_ratings.rating.iloc[index]))
